pyproject.toml
[project]
name = "movesic"
# version = "0.0.0"
dynamic = ["version"]
description = "Simply move your music between services"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    # ui
    "nicegui>=3.4.1",
    "greenlet>=3.3.0",
    # core
    "ytmusicapi>=1.11.3",
    "spotipy>=2.25.2",
    # db
    "sqlalchemy>=2.0.45",
    "alembic>=1.17.2",
    "aiosqlite>=0.22.0",
    # util
    "platformdirs>=4.5.1",
    "lxml>=6.0.2",
    "human-readable>=2.0.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
    "pylance>=1.0.0",
    "pytest>=9.0.2",
    "pytest-alembic>=0.12.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]

[project.optional-dependencies]
native = ["pywebview>=5.1", "pygobject>=3.54.5 ; sys_platform == 'linux'"]

[project.scripts]
movesic = "movesic.__main__:main"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.ruff]
exclude = ["movesic/database/migrations/versions"]

[tool.ruff.format]
preview = true

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"
